Waypoints wiki updates
Clarifications on waypoints:

Uncomment: cleared up the Javascript confusion. This is a HTML comment
waypoint.
Style: added clarification
Font-family: the previous wiki was showing the exact same text as a
different waypoint for importing a google font. Completely overhauled to
reflect the built-in font-family waypoint
Font-size: added what px means and link to other units