Simplified Solution and Page Cleanup
I added a simplified explanation and solution while leaving the original
solution up as an alternate. It is not really necessary to create a
variable to hold the value which is why I put the simplified solution.
For learning purposes though, the original is left up and explained as
an alternate solution. I also reworded the explanation to fit the
solutions.