Enhanced description about "return" statements in JS
While doing the challenges in order, I didn’t encounter the “stop
execution and exit” property of return statements described anywhere.
So, while doing this
[challenge](http://www.freecodecamp.com/challenges/use-conditional-logic
-with-if-statements) I figured that a camper might be confused as to
why the second statement won’t be executed if the condition is true.
Hence, I decided to add that part in the wiki in case someone looks for
it.