Known Issues With Codepen
I spent about a week trying to debug a problem that ended up being an issue with a Chrome extension conflicting with Codepen. This file is meant to provide a list of known issues with Codepen, to reduce frustration in the future.

Rename known-issues-with-codepen to known-issues-with-codepen.md

I added ".md" to the file name. Hopefully this resolved the issue.

Update known-issues-with-codepen.md

Update known-issues-with-codepen.md

Update and rename known-issues-with-codepen.md to known-issue-with-codepen.md

Update known-issue-with-codepen.md

Renamed filename