Create PHP.md
Created PHP page giving brief purpose of the language along with links to the PHP documentation and release notes. Will update this page as I build out the educational pages starting with the basics. I've thrown in a few of the subjects I'll be covering in the near future so that people can see what is to come and I also don't forget what to work on.