Improve clarity of wiki articles, including label descriptions (#862)
* Improve clarity of introductory text and descriptions

Among other things, I attempted to rewrite the descriptions to follow a
consistent structure. Descriptions now *mostly* follow a format along
the lines of, “If this label is applied, it means *the issue blah blah
blah*.” Maybe it looks like bike shedding, but clearer documentation =
better adherence.

* Improve readability and accuracy of the text

Attempted to clarify/correct language and reduce redundancy, to
encourage comprehension and compliance.