Update algorithm truncate (#1091)
* Added advanced solution - algorithm truncate

Added in an advanced one-line solution that uses nested ternary operators. The solution also utilizes the slice() method - which is what the student is "supposed" to use to solve this Algorithm challenge. The current solution uses the substr() method, so I added that into the hint section as well.

* removed nested ternary operators

Removed nested ternary operators and added in new solution with a single ternary operator and slice() method. Updated code description to match single ternary and slice() method.

* Updated to include new template

Added in template layout, and basic solution