Added a check to see if "tracks" exists for the given element of the array.
The existing solution does not allow the camper to pass the challenge. I'm a beginner and am having trouble with JSON more than anything else, so there may be some mistakes in the way I did things. However, this solution that I put here let me pass the challenge, whereas the existing solution did not.