Updated Map wiki page to fix Algorithm-Where-Art-Thou link
The Algorithm-Where-Art-Thou pointed to the wiki page
Algorithm-Where-Art-Thou, which does not exist. The correct wiki page is
Algorithm-Wherefore-Art-Thou, so I updated the text for the link to
match the name of the challenge as it appears on FCC's website and
updated the link as well to point to the existing page.